pyproject.toml
[tool.poetry]
name = "morpheus-api-gateway"
version = "0.1.0"
description = "API Gateway for Morpheus-Lumerin AI Marketplace (FastAPI version)"
authors = ["Your Name <you@example.com>"]
readme = "README.md" # Assuming we will create/update this later
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11" # Or your desired Python version
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
pydantic = {extras = ["email"], version = "^2.6.4"}
pydantic-settings = "^2.2.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.28"}
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9" # For PostgreSQL connection (consider asyncpg for full async)
asyncpg = "^0.29.0" # Preferred async driver for PostgreSQL with SQLAlchemy 2.0+
redis = {extras = ["hiredis"], version = "^5.0.3"}
httpx = "^0.27.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
cryptography = "^42.0.5"
python-dotenv = "^1.0.1"
gunicorn = "^21.2.0"
boto3 = "^1.34.29"  # AWS SDK for Python (for KMS integration)
# Add other dependencies as needed, e.g., slowapi for rate limiting

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5"
mypy = "^1.9.0"
ruff = "^0.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 